// lib/generated/prisma/enums.ts
export const PoiCategory = {
  BEACH: 'BEACH',
  PARK: 'PARK',
  VIEWPOINT: 'VIEWPOINT',
  GOLF_COURSE: 'GOLF_COURSE',
  MARINA: 'MARINA',
  TEMPLE: 'TEMPLE',
  INTERNATIONAL_SCHOOL: 'INTERNATIONAL_SCHOOL',
  LOCAL_SCHOOL: 'LOCAL_SCHOOL',
  KINDERGARTEN: 'KINDERGARTEN',
  UNIVERSITY: 'UNIVERSITY',
  HOSPITAL: 'HOSPITAL',
  CLINIC: 'CLINIC',
  PHARMACY: 'PHARMACY',
  DENTIST: 'DENTIST',
  SHOPPING_MALL: 'SHOPPING_MALL',
  SUPERMARKET: 'SUPERMARKET',
  CONVENIENCE_STORE: 'CONVENIENCE_STORE',
  MARKET: 'MARKET',
  GYM: 'GYM',
  COWORKING: 'COWORKING',
  BANK: 'BANK',
  ATM: 'ATM',
  POST_OFFICE: 'POST_OFFICE',
  RESTAURANT: 'RESTAURANT',
  CAFE: 'CAFE',
  BAR: 'BAR',
  NIGHTCLUB: 'NIGHTCLUB',
  MAIN_ROAD: 'MAIN_ROAD',
  HIGHWAY_ACCESS: 'HIGHWAY_ACCESS',
  AIRPORT: 'AIRPORT',
  BUS_STATION: 'BUS_STATION',
  FERRY_TERMINAL: 'FERRY_TERMINAL',
  TAXI_STAND: 'TAXI_STAND',
  SEA_VIEW_POINT: 'SEA_VIEW_POINT',
  MOUNTAIN_VIEW: 'MOUNTAIN_VIEW',
  RIVER: 'RIVER',
  LAKE: 'LAKE',
  NIGHTLIFE_CLUSTER: 'NIGHTLIFE_CLUSTER',
  BUSY_INTERSECTION: 'BUSY_INTERSECTION',
  CONSTRUCTION_ZONE: 'CONSTRUCTION_ZONE',
  NEIGHBORHOOD: 'NEIGHBORHOOD',
  TOURIST_AREA: 'TOURIST_AREA',
  RESIDENTIAL_AREA: 'RESIDENTIAL_AREA',
  HOTEL: 'HOTEL',
  RESORT: 'RESORT'
} as const

export type PoiCategory = (typeof PoiCategory)[keyof typeof PoiCategory]